# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "轻记 · 本地概念默写",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/imports/preview": {
      "post": {
        "tags": [
          "题库导入"
        ],
        "summary": "Preview Import",
        "operationId": "preview_import_api_imports_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/confirm": {
      "post": {
        "tags": [
          "题库导入"
        ],
        "summary": "Confirm Import",
        "operationId": "confirm_import_api_imports_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportConfirmed"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Read Settings",
        "operationId": "read_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Update Settings",
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/test": {
      "post": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Test Connection",
        "operationId": "test_connection_api_settings_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionFeedback"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/precheck": {
      "post": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Precheck",
        "operationId": "precheck_api_precheck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecheckInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrecheckFeedback"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/error-types": {
      "get": {
        "tags": [
          "设置与核对"
        ],
        "summary": "List Error Types",
        "operationId": "list_error_types_api_error_types_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NameView"
                  },
                  "type": "array",
                  "title": "Response List Error Types Api Error Types Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Create Error Type",
        "operationId": "create_error_type_api_error_types_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NameView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/error-types/{identifier}": {
      "put": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Rename Error Type",
        "operationId": "rename_error_type_api_error_types__identifier__put",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NameView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "设置与核对"
        ],
        "summary": "Delete Error Type",
        "operationId": "delete_error_type_api_error_types__identifier__delete",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConfirmImportRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 10000000,
            "minLength": 1,
            "title": "Content"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Name"
          },
          "duplicate_mode": {
            "type": "string",
            "enum": [
              "skip",
              "overwrite"
            ],
            "title": "Duplicate Mode",
            "default": "skip"
          },
          "preview_token": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Preview Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content",
          "preview_token"
        ],
        "title": "ConfirmImportRequest"
      },
      "ConnectionFeedback": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "title": "ConnectionFeedback"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportConfirmed": {
        "properties": {
          "added": {
            "type": "integer",
            "title": "Added",
            "default": 0
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "default": 0
          },
          "overwritten": {
            "type": "integer",
            "title": "Overwritten",
            "default": 0
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid",
            "default": 0
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "preview_token": {
            "type": "string",
            "title": "Preview Token",
            "default": ""
          },
          "folder_id": {
            "type": "integer",
            "title": "Folder Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rows",
          "folder_id"
        ],
        "title": "ImportConfirmed"
      },
      "ImportReport": {
        "properties": {
          "added": {
            "type": "integer",
            "title": "Added",
            "default": 0
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "default": 0
          },
          "overwritten": {
            "type": "integer",
            "title": "Overwritten",
            "default": 0
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid",
            "default": 0
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "preview_token": {
            "type": "string",
            "title": "Preview Token",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "ImportReport"
      },
      "ImportRequest": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 10000000,
            "minLength": 1,
            "title": "Content"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "folder_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Id"
          },
          "folder_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder Name"
          },
          "duplicate_mode": {
            "type": "string",
            "enum": [
              "skip",
              "overwrite"
            ],
            "title": "Duplicate Mode",
            "default": "skip"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ImportRequest"
      },
      "ImportRow": {
        "properties": {
          "term": {
            "type": "string",
            "title": "Term"
          },
          "definition": {
            "type": "string",
            "title": "Definition"
          },
          "action": {
            "type": "string",
            "enum": [
              "add",
              "skip",
              "overwrite",
              "invalid"
            ],
            "title": "Action"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "term",
          "definition",
          "action",
          "reason"
        ],
        "title": "ImportRow"
      },
...
      "NameInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "NameInput"
      },
      "NameView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "NameView"
      },
...
      "PrecheckFeedback": {
        "properties": {
          "correct_parts": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "title": "Correct Parts"
          },
          "wrong_parts": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "title": "Wrong Parts"
          },
          "uncertain_parts": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "title": "Uncertain Parts"
          },
          "clarifying_questions": {
            "items": {
              "type": "string",
              "minLength": 1
            },
            "type": "array",
            "maxItems": 3,
            "title": "Clarifying Questions"
          },
          "suggested_rewrite": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Rewrite"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "correct_parts",
          "wrong_parts",
          "uncertain_parts",
          "clarifying_questions",
          "suggested_rewrite"
        ],
        "title": "PrecheckFeedback"
      },
      "PrecheckInput": {
        "properties": {
          "term": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Term"
          },
          "definition": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Definition"
          },
          "reference_note": {
            "type": "string",
            "maxLength": 100000,
            "title": "Reference Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "term",
          "definition"
        ],
        "title": "PrecheckInput"
      },
...
      "SettingsInput": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "deepseek",
                  "gemini",
                  "openai",
                  "kimi",
                  "qwen",
                  "custom"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "judge_count": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  2,
                  3
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Count"
          },
          "judge_temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Temperature"
          },
          "merge_temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merge Temperature"
          },
          "w_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "W Accuracy"
          },
          "w_completeness": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "W Completeness"
          },
          "pass_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pass Threshold"
          },
          "json_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "auto",
                  "on",
                  "off"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Json Mode"
          },
          "request_timeout": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Timeout"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SettingsInput"
      },
      "SettingsView": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "judge_count": {
            "type": "integer",
            "title": "Judge Count"
          },
          "judge_temperature": {
            "type": "number",
            "title": "Judge Temperature"
          },
          "merge_temperature": {
            "type": "number",
            "title": "Merge Temperature"
          },
          "w_accuracy": {
            "type": "number",
            "title": "W Accuracy"
          },
          "w_completeness": {
            "type": "number",
            "title": "W Completeness"
          },
          "pass_threshold": {
            "type": "number",
            "title": "Pass Threshold"
          },
          "json_mode": {
            "type": "string",
            "title": "Json Mode"
          },
          "request_timeout": {
            "type": "number",
            "title": "Request Timeout"
          },
          "has_api_key": {
            "type": "boolean",
            "title": "Has Api Key"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "base_url",
          "model",
          "judge_count",
          "judge_temperature",
          "merge_temperature",
          "w_accuracy",
          "w_completeness",
          "pass_threshold",
          "json_mode",
          "request_timeout",
          "has_api_key"
        ],
        "title": "SettingsView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}